api/requirements.txt
fastapi
uvicorn
pydantic
transformers
torch
psycopg2-binary 
python-dotenv 
python-multipart
fastapi[standard]